python-analyzer/requirements.txt
python-chess==1.999
stockfish==3.28.0
pytest==8.1.1
requests==2.31.0
